Refactor dashboard operator auth state
